Preliminary work on a GAF loader. This will likely change in the very near future.
The issue with this system is that it expands the 8-bit palette values to an RGBA
value based on the palette which in turn results in a lot of wasted memory. Future
commits will be experimenting with using a pixel shader to implement a palette
based texturing system rather than a true color RGBA texturing.